python-parser/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
pdfplumber>=0.11.4
camelot-py>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.13
Pillow>=11.0.0
SQLAlchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.9.2
